SCTASK0150694
Removing the updated file in the format of Press release. Matt mentioned to John that it is not a press release we need to update the text.

On January 30, 2020, Patriot Transportation Holding, Inc. (the "Company") paid a cash dividend of $3.15 per share to shareholders of record at the close of business on January 15, 2020. 
 
On December 4, 2020, the Board of directors declared a special cash dividend of $3.00 per share, payable on December 30, 2020 to the Company’s shareholders of record as of December 17, 2020.  The ex-dividend date for this dividend is December 31, 2020.
 
Shareholders should consult with their individual tax advisors regarding these dividends. Patriot has sufficient accumulated earnings and profits for both dividends to qualify as a dividend for tax purposes rather than a return of capital. As Patriot is a U.S. corporation, Patriot management believes that both special dividends will qualify as a “qualified dividend” for shareholders who meet the requisite holding period.