Testing Zoom Meeting
From: Wagner, John L 
Sent: Wednesday, January 20, 2021 6:00 PM
To: Ashbrook, Roy <RAshbrook@patriottrans.com>; Mahajan, Pranav <pmahajan@patriottrans.com>
Subject: Test Zoom Link on Investor Site

We need to test the publishing and use of a Zoom link on the Investor site for the upcoming board meeting.  Matt wants to use a Zoom meeting and wants he and Rob to be able to click a link on the Investor page to access the meeting.  They want the attendees to be able to hear and see a presented document, but not have video access.  Please publish a test link on this page (http://patriottrans.com/investor-relations.html) and let me know when its ready to test.  Please work on this tomorrow.  The meeting is set for 2/3 and Matt wants the link to the real meeting up “soon”.

Let me know what additional info you need.  Thanks.


John Wagner
CIO, Patriot Transportation